refactor: use observability package in server and mcp
