+++ apps/server-mcp/package.json
@@ -15,14 +15,10 @@
   "dependencies": {
     "@effect/ai": "^0.33.2",
     "@effect/experimental": "^0.58.0",
-    "@effect/opentelemetry": "^0.61.0",
     "@effect/platform": "^0.94.5",
     "@effect/platform-bun": "^0.87.1",
-    "@opentelemetry/api": "^1.9.0",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.212.0",
-    "@opentelemetry/sdk-trace-base": "^2.5.1",
-    "@opentelemetry/sdk-trace-node": "^2.5.1",
     "@repo/domain": "workspace:*",
+    "@repo/observability": "workspace:*",
     "effect": "3.19.19"
   },
   "devDependencies": {
