+++ apps/server-mcp/src/index.ts
@@ -1,11 +1,9 @@
 import { McpServer, Tool, Toolkit } from "@effect/ai";
 import { DevTools } from "@effect/experimental";
-import { NodeSdk } from "@effect/opentelemetry";
 import { HttpLayerRouter, HttpServer } from "@effect/platform";
 import { BunHttpServer, BunRuntime } from "@effect/platform-bun";
-import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-http";
-import { BatchSpanProcessor } from "@opentelemetry/sdk-trace-base";
-import { Config, Effect, Layer, Option, Schema } from "effect";
+import { ObservabilityLive } from "@repo/observability";
+import { Config, Effect, Layer, Schema } from "effect";
 
 // Define Resources
 const ResourceLayer = Layer.mergeAll(
@@ -73,11 +71,6 @@ const ServerConfig = Config.all({
   enableDevTools: Config.boolean("DEVTOOLS").pipe(Config.withDefault(false)),
 });
 
-const TracingConfig = Config.all({
-  exporterEndpoint: Config.option(Config.string("OTEL_EXPORTER_OTLP_ENDPOINT")),
-  serviceName: Config.option(Config.string("OTEL_SERVICE_NAME")),
-});
-
 const McpRouter = McpServer.layerHttpRouter({
   name: "BEVR MCP Server",
   version: "0.1.0",
@@ -94,27 +87,6 @@ const McpRouter = McpServer.layerHttpRouter({
   ),
 );
 
-const NodeSdkLive = Effect.gen(function* () {
-  const tracing = yield* TracingConfig;
-  const endpoint = Option.getOrUndefined(tracing.exporterEndpoint);
-  const serviceName = Option.getOrUndefined(tracing.serviceName);
-
-  if (!endpoint || !serviceName) {
-    yield* Effect.log(
-      "OTEL tracing disabled (set OTEL_EXPORTER_OTLP_ENDPOINT and OTEL_SERVICE_NAME to enable)",
-    );
-    return Layer.empty;
-  }
-
-  yield* Effect.log(`OTEL tracing enabled: ${serviceName} -> ${endpoint}`);
-  return NodeSdk.layer(() => ({
-    resource: { serviceName },
-    spanProcessor: new BatchSpanProcessor(
-      new OTLPTraceExporter({ url: endpoint }),
-    ),
-  }));
-}).pipe(Layer.unwrapEffect);
-
 const DevToolsLive = Effect.gen(function* () {
   const config = yield* ServerConfig;
   if (!config.enableDevTools) {
@@ -127,7 +99,7 @@ const DevToolsLive = Effect.gen(function* () {
 const HttpLive = HttpLayerRouter.serve(McpRouter).pipe(
   HttpServer.withLogAddress,
   Layer.provideMerge(DevToolsLive),
-  Layer.provideMerge(NodeSdkLive),
+  Layer.provideMerge(ObservabilityLive),
   Layer.provideMerge(BunHttpServer.layerConfig(ServerConfig)),
 );
