+++ packages/observability/src/index.ts
@@ -1,0 +1,32 @@
+import { NodeSdk } from "@effect/opentelemetry";
+import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-http";
+import { BatchSpanProcessor } from "@opentelemetry/sdk-trace-base";
+import { Config, Effect, Layer, Option } from "effect";
+
+const TracingConfig = Config.all({
+  exporterEndpoint: Config.option(Config.string("OTEL_EXPORTER_OTLP_ENDPOINT")),
+  serviceName: Config.option(Config.string("OTEL_SERVICE_NAME")),
+});
+
+export const Observability = NodeSdk;
+
+export const ObservabilityLive = Effect.gen(function* () {
+  const tracing = yield* TracingConfig;
+  const endpoint = Option.getOrUndefined(tracing.exporterEndpoint);
+  const serviceName = Option.getOrUndefined(tracing.serviceName);
+
+  if (!endpoint || !serviceName) {
+    yield* Effect.log(
+      "OTEL tracing disabled (set OTEL_EXPORTER_OTLP_ENDPOINT and OTEL_SERVICE_NAME to enable)",
+    );
+    return Layer.empty;
+  }
+
+  yield* Effect.log(`OTEL tracing enabled: ${serviceName} -> ${endpoint}`);
+  return NodeSdk.layer(() => ({
+    resource: { serviceName },
+    spanProcessor: new BatchSpanProcessor(
+      new OTLPTraceExporter({ url: endpoint }),
+    ),
+  }));
+}).pipe(Layer.unwrapEffect);
